requirements.txt
langgraph
langchain
langchain-core
langchain_community
langchain_huggingface
langchain-ollama
python-dotenv
selenium
webdriver_manager
bs4
langchain-ollama
requests
torch 